Reorganize routes and add collection fn to model.
